[RSDK-11346, RSDK-11355] Split out main.cpp into separate files (#27)
* wip

* Add new file in cmakelists; Remove curly brace semicolon; Make service serve call a new loc; Rename state_ to props_ and make it public (move to hpp); Use orbbec namespace instead of anonymous namespace in orbbec.cpp; Don't alias vsdk in hpp file as that affects downstream

* resource_properties_ -> resource_properties

* resource_properties -> ObResourceConfig

* Unhardcode string data format consts; Add nil checks before serial number logs in destructor; Add nil checks for color and depth data; Add comment about native config

* Use abbrev viam::sdk; Remove struct keyword
